msm: mdss: Add a mdss_dsi_is_panel_dead() api
The panel_dead flag will be set when the check_status thread detects the
ESD event from the panel which is connected to DSI0.

Since QHD panel is a dual-link panel which will be connected to both
DSI0 and DSI1 links. The DSI0 can check the panel_dead status in its data
structure and act accordingly, but DSI1 can not. This new API will
return the panel_dead flag status from DSI0 to DSI1.

Change-Id: I8a11d880b46b3b41fd9fe752131b5b8c152caa1a
Signed-off-by: Bang Nguyen <bnguyen1@motorola.com>
Reviewed-by: Joseph Swantek <jswantek@motorola.com>
Reviewed-by: Zachary Crosby <zac@motorola.com>
Reviewed-by: Stephen Rossbach <rossbach@motorola.com>