misc: lis3dh_mot: Fix CTRL_REG4 configuration when enabling IRQ
BLK_UPDATE is equal to 0x80, so the content of CTRL_REG4 is always set
to either 0x80 or 0x00 when the IRQ is enabled. Most likely the purpose
of the code was setting the threshold to G_2G while making that BDU was
set to 1, so change it to make this happen.

In addition to that, the value of resume_state[3] was set equal to
the content of CTRL_REG5 when it actually should contains the content
of CTRL_REG4. So move the assignement so that resume_state[3] contains
the correct value.

Change-Id: Icca6db20a990608da9f3c6ef3ee50acd0d171ff9